refactor: repository 패키지 생성
