+++ back-end/src/main/java/com/tdd/backend/post/service/SharingService.java
@@ -2,7 +2,7 @@ package com.tdd.backend.post.service;
 
 import org.springframework.stereotype.Service;
 
-import com.tdd.backend.post.PostRepository;
+import com.tdd.backend.post.repository.PostRepository;
 import com.tdd.backend.post.data.SharingDto;
 
 import lombok.RequiredArgsConstructor;
